feat(article): 重构文章模块并添加新功能
- 重构了 ArticleController，优化了文章管理相关的接口
- 新增了 ArticleQueryVO 类，用于文章查询
- 在 ArticleEntity 中添加了额外的字段，如 cateIds、tagIds 等
- 添加了 AmapLocationUtil 工具类，用于获取地理位置信息
- 更新了 application.yml 配置文件，增加了文件存储相关配置

+++ src/main/java/top/luoyuanxiang/thrivex/server/entity/ArticleConfigEntity.java
@@ -44,19 +44,19 @@ public class ArticleConfigEntity extends Model<ArticleConfigEntity> {
      * 是否加密
      */
     @TableField("is_encrypt")
-    private Byte isEncrypt;
+    private Integer isEncrypt;
 
     /**
      * 是否为草稿
      */
     @TableField("is_draft")
-    private Byte isDraft;
+    private Integer isDraft;
 
     /**
      * 是否删除
      */
     @TableField("is_del")
-    private Byte isDel;
+    private Integer isDel;
 
     /**
      * 对应的文章id
